+++ packages/porygon/CHANGELOG.md
@@ -20,3 +20,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [UNRELEASED]
+
+### Added
+- Added `Interactor` model.
+- Added `InteractorAsync` model.
+- Added `CreateEntityPort` model.
+- Added `DeleteEntityPort` model.
+- Added `ReadEntityPort` model.
+- Added `UpdateEntityPort` model.
+- Added `ModuleTypeToSymbolMap` model.
+- Added `CrudModuleType` enum.
+- Added `DomainCrudContainerModuleApi` module.